+++ src/main/kotlin/blackjack/core/player/Players.kt
@@ -1,4 +1,4 @@
-package blackjack.core
+package blackjack.core.player
 
 class Players(val players: List<Player>) : List<Player> by players {
     constructor(names: Set<Name>) : this(initPlayers(names))
